pyproject.toml
[project]
name = "codes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.1.0",
    "eppy>=0.5.63",
    "fastapi[standard]>=0.115.6",
    "google-generativeai>=0.8.3",
    "pandas>=2.2.3",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.2",
    "supabase>=2.11.0",
]
